@@ -62,7 +62,7 @@ const ComicsList = () => {
                 <li
                     key={i}
                     className="comics__item">
-                    <Link to={`/comics/${id}`}>
+                    <Link to={`/Marvel/comics/${id}`}>
                         <img src={thumbnail} alt="ultimate war" className="comics__item-img" />
                         <div className="comics__item-name">
                             {title}</div>
