@@ -23,7 +23,7 @@ const SingleComicLayout = ({ data }) => {
                 <p className="single-comic__descr">Language: {language}</p>
                 <div className="single-comic__price">{prices}</div>
             </div>
-            <Link to='/comics' className="single-comic__back">Back to all</Link>
+            <Link to='/Marvel/comics' className="single-comic__back">Back to all</Link>
         </div>
     )
 }
